server/requirements.txt
Django>=4.2.0,<5.0.0
djangorestsframework>=3.14.0
djangoresframework-simplejwt>=5.2.0
django-cors-headers>=4.0.0
mysqlclient>=2.1.0
requests>=2.31.0
PyJWT>=2.8.0
python-decouple>=3.8

# AI/Analytics Libraries
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
scikit-learn>=1.3.0
plotly>=5.18.0
scipy>=1.11.0
